+++ src/main.hpp
@@ -107,6 +107,8 @@ using std::string; //Instead of including an entire namespace, please explicitly
 #define access _access
 #endif
 
+#include "list.hpp"
+
 #define PI 3.14159265358979323846
 
 extern FILE* logfile;
@@ -232,22 +234,8 @@ extern view_t camera;
 class Entity; //TODO: Bugger?
 
 // node structure
-typedef struct node_t
-{
-	struct node_t* next;
-	struct node_t* prev;
-	struct list_t* list;
-	void* element;
-	void (*deconstructor)(void* data);
-	Uint32 size;
-} node_t;
 
 // list structure
-typedef struct list_t
-{
-	node_t* first;
-	node_t* last;
-} list_t;
 extern list_t button_l;
 extern list_t light_l;
 
@@ -519,18 +507,6 @@ int numdigits_sint16(Sint16 x);
 int longestline(char* str);
 int concatedStringLength(char* str, ...);
 void printlog(const char* str, ...);
-
-// function prototypes for list.c:
-void list_FreeAll(list_t* list);
-void list_RemoveNode(node_t* node);
-node_t* list_AddNodeFirst(list_t* list);
-node_t* list_AddNodeLast(list_t* list);
-node_t* list_AddNode(list_t* list, int index);
-Uint32 list_Size(list_t* list);
-list_t* list_Copy(list_t* destlist, list_t* srclist);
-list_t* list_CopyNew(list_t* srclist);
-Uint32 list_Index(node_t* node);
-node_t* list_Node(list_t* list, int index);
 
 // function prototypes for light.c:
 light_t* lightSphereShadow(Sint32 x, Sint32 y, Sint32 radius, Sint32 intensity);
